Added tooling and changed code to TypeScript
* Installed TypeScript and related rules for ESLint
* Modified ESLint and Prettier rules to work with TS
* Added ignores to git and eslint
* Updated README.md
