+++ app.json
@@ -1,4 +1,4 @@
 {
   "name": "cooky",
-  "displayName": "cooky"
+  "displayName": "Cooky"
 }
