+++ src/App.tsx
@@ -1,0 +1,91 @@
+import React from 'react';
+import { SafeAreaView, ScrollView, StatusBar, StyleSheet, Text, View } from 'react-native';
+
+import {
+    DebugInstructions,
+    Header,
+    LearnMoreLinks,
+    ReloadInstructions,
+} from 'react-native/Libraries/NewAppScreen';
+
+const App: any = () => {
+    return (
+        <>
+            <StatusBar barStyle="dark-content" />
+            <SafeAreaView>
+                <ScrollView contentInsetAdjustmentBehavior="automatic" style={styles.scrollView}>
+                    <Header />
+                    <View style={styles.body}>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>Step One</Text>
+                            <Text style={styles.sectionDescription}>
+                                Edit <Text style={styles.highlight}>App.js</Text> to change this
+                                screen and then come back to see your edits.
+                            </Text>
+                        </View>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>See Your Changes</Text>
+                            <Text style={styles.sectionDescription}>
+                                <ReloadInstructions />
+                            </Text>
+                        </View>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>Debug</Text>
+                            <Text style={styles.sectionDescription}>
+                                <DebugInstructions />
+                            </Text>
+                        </View>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>Learn More</Text>
+                            <Text style={styles.sectionDescription}>
+                                Read the docs to discover what to do next:
+                            </Text>
+                        </View>
+                        <LearnMoreLinks />
+                    </View>
+                </ScrollView>
+            </SafeAreaView>
+        </>
+    );
+};
+
+const styles = StyleSheet.create({
+    scrollView: {
+        backgroundColor: '#888',
+    },
+    engine: {
+        position: 'absolute',
+        right: 0,
+    },
+    body: {
+        backgroundColor: 'white',
+    },
+    sectionContainer: {
+        marginTop: 32,
+        paddingHorizontal: 24,
+    },
+    sectionTitle: {
+        fontSize: 24,
+        fontWeight: '600',
+        color: 'black',
+    },
+    sectionDescription: {
+        marginTop: 8,
+        fontSize: 18,
+        fontWeight: '400',
+        color: '#222',
+    },
+    highlight: {
+        fontWeight: '700',
+    },
+    footer: {
+        color: '#222',
+        fontSize: 12,
+        fontWeight: '600',
+        padding: 4,
+        paddingRight: 12,
+        textAlign: 'right',
+    },
+});
+
+export default App;
